// src/models/brands/types/BrandBelongsToCompanyRelationship.ts
import type {BaseNode} from "../../../controllers/nodes/types/BaseNode"

export type BrandBelongsToCompanyRelationship = {
    brand_id: number
    company_id: number
    relationship_id: number
    relationship_name: string
    relationship_partner?: BaseNode
    created_at: string
    updated_at: string
}